+++ lib/src/cook/process/souschef/Archiver.cpp
@@ -30,13 +30,12 @@ Result Archiver::process(model::Recipe & recipe, RecipeFilteredGraph & file_comm
     {
         object.set_propagation(Propagation::Private);
         auto ss = log::scope("object", [&](auto &node){node.attr("file", object);});
-        const std::filesystem::path & obj_fn = util::make_global_from_recipe(recipe, object.key());
-
+        const std::filesystem::path & obj_fn = gubg::filesystem::combine({ recipe.working_directory(), object.dir(), object.rel() });
         MSS(g.add_edge(archive_vertex, g.goc_vertex(obj_fn)));
     }
 
     // get the library dir (local to the path)
-    const std::filesystem::path & library_dir = util::make_local_to_recipe(util::make_recipe_adj_path(recipe), context.dirs().output());
+    std::filesystem::path library_dir = util::get_from_to_path(recipe, context.dirs().output(true));
     {
         MSS(!!recipe.build_target().filename);
         ingredient::File dep(library_dir, *recipe.build_target().filename);
@@ -49,7 +48,7 @@ Result Archiver::process(model::Recipe & recipe, RecipeFilteredGraph & file_comm
 
         // add the link in the execution graph
         {
-            const std::filesystem::path & lib_fn = util::make_global_from_recipe(recipe, dep.key());
+            const std::filesystem::path & lib_fn = context.dirs().output(true) / dep.key();
             MSS(g.add_edge(g.goc_vertex(lib_fn), archive_vertex));
         }
     }