+++ lib/src/cook/process/souschef/Compiler.hpp
@@ -16,7 +16,7 @@ public:
     Result process(model::Recipe & recipe, RecipeFilteredGraph & file_command_graph, const Context & context) const override;
 
 private:
-    ingredient::File construct_object_file(const ingredient::File & source, model::Recipe &recipe, const Context &context, const std::filesystem::path &adj_path) const;
+    ingredient::File construct_object_file(const ingredient::File & source, model::Recipe &recipe, const Context &context) const;
     Result compile_command_(command::Ptr &, const model::Recipe &recipe, const Context & context) const;
 
     Language language_;