+++ lib/src/cook/process/souschef/Linker.cpp
@@ -51,25 +51,25 @@ Result Linker::process(model::Recipe & recipe, RecipeFilteredGraph & file_comman
     for(ingredient::File & object : objects)
     {
         auto ss = log::scope("object", [&](auto &node){node.attr("file", object);});
-        const std::filesystem::path & obj_fn = util::make_global_from_recipe(recipe, object.key());
+        const std::filesystem::path & obj_fn = gubg::filesystem::combine({recipe.working_directory(), object.dir(), object.rel()});
         MSS(g.add_edge(link_vertex, g.goc_vertex(obj_fn)));
     }
 
     // link the dependencies
     for(ingredient::File & dep : deps)
     {
-        const std::filesystem::path & fn = util::make_global_from_recipe(recipe, dep.key());
+        const std::filesystem::path & fn = gubg::filesystem::combine({recipe.working_directory(), dep.dir(), dep.rel()});
         MSS(g.add_edge(link_vertex, g.goc_vertex(fn), RecipeFilteredGraph::Implicit));
     }
     for(ingredient::File & exp : exports)
     {
-        const std::filesystem::path & fn = util::make_global_from_recipe(recipe, exp.key());
+        const std::filesystem::path & fn = gubg::filesystem::combine({recipe.working_directory(), exp.dir(), exp.rel()});
         MSS(g.add_edge(link_vertex, g.goc_vertex(fn), RecipeFilteredGraph::Implicit));
     }
 
 
     // create the local link dir
-    const std::filesystem::path & library_dir = util::make_local_to_recipe(util::make_recipe_adj_path(recipe), context.dirs().output());
+    const std::filesystem::path & library_dir = util::get_from_to_path(recipe, context.dirs().output(true));
 
     // create the linked file
     {
@@ -103,7 +103,7 @@ Result Linker::process(model::Recipe & recipe, RecipeFilteredGraph & file_comman
 
         // add the link in the execution graph
         {
-            const std::filesystem::path & lib_fn = util::make_global_from_recipe(recipe, library_dir) / *recipe.build_target().filename;
+            const std::filesystem::path & lib_fn = context.dirs().output(true) / *recipe.build_target().filename;
             MSS(g.add_edge(g.goc_vertex(lib_fn), link_vertex));
         }
 
@@ -162,6 +162,8 @@ Result Linker::link_command_(command::Ptr &ptr, const model::Recipe & recipe, co
     auto lp = element->create<process::command::Link>();
     MSS(!!lp);
 
+    auto adj = util::get_from_to_path(".", recipe);
+
     // set the libraries
     for(const ingredient::KeyValue & lib : libs)
         lp->add_library(lib.key());
@@ -172,15 +174,15 @@ Result Linker::link_command_(command::Ptr &ptr, const model::Recipe & recipe, co
 
     // set the library paths
     for(const ingredient::File & lib: recipe.files().range(LanguageTypePair(Language::Binary, Type::LibraryPath)))
-        lp->add_library_path(util::make_global_from_recipe(recipe, lib.dir()));
+        lp->add_library_path(gubg::filesystem::combine(adj, lib.dir()));
 
-    // set the frameowork paths
+    // set the framework paths
     for(const ingredient::File & framework_path: recipe.files().range(LanguageTypePair(Language::Binary, Type::FrameworkPath)))
-        lp->add_framework_path(util::make_global_from_recipe(recipe, framework_path.dir()));
+        lp->add_framework_path(gubg::filesystem::combine(adj, framework_path.dir()));
 
     // set the export files
     for(const ingredient::File & exp: recipe.files().range(LanguageTypePair(Language::Definition, Type::Export)))
-        lp->add_export(util::make_global_from_recipe(recipe, exp.key()).string());
+        lp->add_export(gubg::filesystem::combine({adj, exp.dir(), exp.rel()}));
 
     lp->set_recipe_uri(recipe.uri().string());
     ptr = lp;